Added API's "from" and "to" parameters
Since the script always begins with the most recent tracks it is not very useful if one wants to get only the latest scrobbles without all the old tracks one might already have. For that the parameters "from" and "to" (in UNIX timestamp) are necessary, so I have added them as options. By default, they are empty (default=None), in which case the script works as before.
A remark for future development:
lastexport.py:132: DeprecationWarning: This method will be removed in future versions. Use 'list(elem)' or iteration over elem instead.
if trackelement.find('artist').getchildren():